+++ doc/sphinx/changelog.rst
@@ -1,6 +1,26 @@
 Changelog
 =========
 
+0.5 (unreleased)
+----------------
+
+New
+~~~
+
+- Add the possibility of constructing non-canonical :cpp:class:`~mppp::rational` objects from numerator/denominator pairs
+  (`#28 <https://github.com/bluescarni/mppp/pull/28>`__).
+
+Changes
+~~~~~~~
+
+- Performance improvements and simplifications in the :cpp:class:`~mppp::rational` constructors
+  (`#28 <https://github.com/bluescarni/mppp/pull/28>`__).
+
+Fix
+~~~
+
+- Fix some MSVC warnings when compiling the tests in release mode (`#28 <https://github.com/bluescarni/mppp/pull/28>`__).
+
 0.4 (2017-07-29)
 ----------------